Adds 'duplicate' available actions
I have a program that has a 'recipe' with 'ingredients' as the child. I wanted to be able to duplicate the 'recipe' and at the same time duplicate all the child rows.

It works on my end. I looked at how you got the primary key of a row, and then added a function 'get_child_pk'.