more work on prompt_save=True parameter for all of the record selection methods, sinch by default we want to prompt for changes to be saved (if enabled).  There are some times however, like on a requery where you want to select the first record afterwards that you don't want to prompt, since for that bit of time the Query data and the PySimpleGUI data will be (understandably) out of sync until the controls update.
Aside from that, giving the user ultimate control over whether prompt_save occurs makes the most sense.

Still need to update the documentation that the prompt_save parameter does not supercede the actual set_prompt_save on either the Query or Form level - it only allows (or not) for the prompt_save action to execute.  This may be a bit confusing, perhaps a better parameter name should be used to avoid confusion.  In any case, this seems to be a good working example to build off of.