refs #56 Since we ran the transform only on the changed items, we actually don't even need to transform back.
Leaving update_elements() for now.  Looking into the case where a quick editor is opened and one of the values changed.  Need to make sure that comboboxes requery with this new value.  For example, open the journal_with_data_manipulation example and use the quick editor to rename the mood from Happy to Happy2.  You will see that the combobox does not get the fresh data since we haven't requeried.  Will be looking into this next