Adds parent_cascade_fk check to save_record
Your new transform code fixed the IndexError when a child changes its parent. It now just updates on screen, until the table requeries... dropping out of view at that time.

This checks if a cascade-fk has changed, and requeries the table, moving the index to the previous in the table.

Also small fix for def previous, so that it doesn't move back two rows.