Explicitly close database
On my system, the -wal -shm files were being left open.

I did some reading, and people say that __del__ can't be completely relied on to ever be called.

Sure enough, if I put a print() in there, it wasn't. Even if I did `del frm`

So I changed the __del__ to db_close, and, added it to close(), and will be replacing:

`frm = None' with `frm.close()`

Let me know, and I can also change the examples