Postgres proper next_pk
This was a head scratcher.  To get this to work with case sensitive table names, the sequence name has to be doulbe quoted, then single quoted around that. For example, with a table Journal and a primary key of id, the query ends up looking like:
SELECT nextval('"Journal_id_seq"');

This has more to do with the choice for pysimplesql to support case-sensitive table names for Postgres so that it works consistently with the other databases (typically, Postgres encourages use of lower-case table names).

In any case, this effectively checks out the primary key, which prevents collisions even with concurrent access.