Defaults is overwriting r.fk_column on insert
I'm not sure if this was intentional or not.

This was using the database default generator, instead of the parent.pk

eg
`Did not set default for id. Setting default for bike_repair_id to 1`

I think in general if inserting a child under an parent, we'd want to pass in the parent pk.