refs #74 Marking notnull records
I've spent way too much time on this.  I can't for the life of me figure out why the asterisks are appearing AFTER the column field, and not before as they are clearly defined.  I'm thinking this has to be a PySimpleGUI bug of some sort.  The sg.Text() element is defined in the layout before the actual element, but displays after the element in the GUI.  I have purposely set the visibility to Tue on creation and you can see a very quick blip on loading where the "marker" appears before the sg.element - but everythin after shows it after the element