refs #102 Column sorting cleanup
Column sorting is getting more integrated and much cleaner now.  All sorting happens at the ResultRow level, the same update_elements is responsible for updating the table rows in all cases.

Made a simple callback wrapper for the TableHeadings callback for when a record changes.  This helps separate logic between ResultSet and TableHeading