Added a Form.set_force_save() method
While working on the Flatfile example, it became apparent that being able to force saves for unchanged records added a lot of utility.  For example, the data could be loaded, the headers clicked to re-order the data, then the save button would write the re-ordered data back to the csv file without the need to actually change a record.