working on MySQL/PostgreSQL examples
MySQL now working.  However, there does appear to be a bug in sort_reset.
If you open the example, you can click on headings and sort with no issues.  However, after editing and saving a record, there is a TypeError: '<' not supported between instances of 'int' and 'NoneType'

Going to try to track this down