Postgres examples catchup and cleanup
Looks to be working now - though the pysimplesql_user is not set up correctly with HelioHost - just waiting for them to complete the trouble ticket.  I did test with the admin credentials and it worked.

Also created some credentials dicts in the MAIN section to clean up the examples a little.