Fix issue with single-file pysimplesql import / Move additional themepacks to theme_packs
If you have pysimplesql in same directory as script, single-file imports would fail with an ImportError.

Additional, in sqldriver, it looks for 'common', but the single-file import only provides 'all'. So I added all common, and renamed the single-file import RESERVED dict to 'common'.

I also move the additional themepacks to theme_pack.py. Seemed to be in line with how we did the language_pack.py