Cleaning up prompt_save, autosave
Per PEP8, put leading _ is for weak internal use, but really we are using it to avoid clashing with function name. So renamed variables to prompt_save_.

Added prompt_save_ to Form, so we can pass down initially set prompt_save_ to Datasets.

Renamed autosave to prompt_silent. This makes it more clear that it is a mode of prompt_save. Passed initially set prompt_silent to Datasets. Added it to set_prompt_save