Postgres fixes
Fixed pysimplesql_user privileges and documented them.  Added exception handling to execute().  A setval() on the sequence was failing and causing future queries to fail due to the failed statement not being rolled back.  Still need to see why setval() is failing when the user has the proper sequence permission, but defaulted the syncing in the constructor until I have more time to sit and look at it