refs #225 MySQL Docker now working
Still some small things to do, but the docker image now works for testing MySQL.

Had to change a few small things with getting table names, as some versions of MySQL return the table_name field as upper case and others as lower case.  Another issue was with the sql data type (domain) is sometimes a bytes object that must be decoded, and other a string - depending on mysql version?