refs # 244 MS Access first minimally working example
Fixed the relationships - names of tables and columns were not being returned in their original case.

I also realized that I rebased off of a local development which was not current with remote, which is why some changes appear lost and ruff is yelling at me.  After this commit, I'll copy the development over by hand and add in the changes by again