refs # 244 MS Access fixes
navigation and saving now working.  There is still a lot of work to do - dates specifically are a little complicated.  In Access, everything is a timestamp, and there is another field to define the type of timestamp - this will need captured at some point.

There is a bug on inserts now, i will work on that next