refs #278 Unit Test example for ProgressAnimate
Not perfect yet, but shows the basic concepts of testing for both no failures and expected failures.  I tried to fix things in ProgressAnimate as I went along so that hopefully you can see the correlation between what was being tested, and the fixes that went in to pass the tests.

If you haven't used pytest before, there's a few things to do:
- pip install pytest
- make sure pysimplesql is reachable as a module.  The easiest way is to do a local installation by running `pip -e .` in the pysimplesql root directory
- from the commandline, run `pytest` to run all tests.  Or you can run tests for a specific file, I.e. `pytest progressanimate_test.py`